make IRSuite test work again
